+++ packages/utils/src/mergeDefaultsWithFormData.ts
@@ -41,7 +41,7 @@ export default function mergeDefaultsWithFormData<T = any>(
     const overrideOppositeArray = overrideFormDataWithDefaults ? formData : defaultsArray;
 
     const mapped = overrideArray.map((value, idx) => {
-      if (overrideOppositeArray[idx]) {
+      if (typeof overrideOppositeArray[idx] !== 'undefined') {
         return mergeDefaultsWithFormData<any>(
           defaultsArray[idx],
           formData[idx],